+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## Unreleased
 
+- Fix summary is not correctly serialized with JSON or YAML output format [#116](https://github.com/BernieWhite/PSRule/issues/116)
 - Fix missing properties on serialized YAML output [#115](https://github.com/BernieWhite/PSRule/issues/115)
 - Fix incorrect property name case of YAML serialized results [#114](https://github.com/BernieWhite/PSRule/issues/114)